fix:GradationBarBox 로 다른 컴포넌트로 분리
